Final Project (3)
Project Description: The intent is to develop a Node.js Web Application to compute the final grade for CSC404. Before the web application was made, project 1 focused on making a prototype for the Web application by using the “prompt” module and prompts users to inquire data. Project 2 takes project 1 into the next step by using node.js to create a working web application. Finally, project 3 continues to go further by replacing the array holding test case information with a database, MongoDB and mongoose. The course is based on three homework assignments with the weight of 20%, and two exams weighted 40% each. The program we implemented calculated the final grade letter as well based on the rubric of A(93 or higher), A-(90-92), B+(87-89), B(83-86), B-(80-82), C-, C, C-, D+, D, D-, and F(0-59). The grade is rounded up if the decimal point is at its border (89.5 becomes 90). Multiple functions were created for the testers on our team to find bugs for each module. The web application includes a layout with three consistent buttons that route to the home page, add test form page, and a results page. The add test page contains a form where users can input their grades for each test and homework. Each test case added will return the final grade percentage and a grade letter for each test case entered in the add test form. Nodemon is used in this project to run the program.